move function to querybuilder interface
since both sqla and django backends now use the 'uuid' column type for
uuids, the get_filter_expr_from_column function becomes identical
for the django and sqla querybuilder, and has been moved to the
querybuilder interface.
